fixed bug that wasn't allowing decimal places in the currencyFormatter function. Fix was switch a parseInt to a parseFloat

+++ server/controllers/currencyFormatter.js
@@ -2,7 +2,7 @@ const currencyFormatter = (currency) => {
   if (!parseInt(currency) && !parseFloat(currency)) {
     return false;
   }
-  return parseInt(currency).toFixed(2);
+  return parseFloat(currency).toFixed(2);
 };
 
 module.exports = currencyFormatter;
